+++ handler.go
@@ -1,10 +1,11 @@
 package main
 
 import (
+	"errors"
 	"fmt"
 	"html/template"
 	"net/http"
-	"sort"
+	"path"
 	"strings"
 
 	"go.uber.org/sally/templates"
@@ -17,112 +18,131 @@ var (
 		template.New("package.html").Parse(templates.Package))
 )
 
-// CreateHandler creates a Sally http.Handler
-func CreateHandler(config *Config) http.Handler {
-	mux := http.NewServeMux()
+// Handler handles inbound HTTP requests.
+//
+// It provides the following endpoints:
+//
+//	GET /
+//		Index page listing all packages.
+//	GET /<name>
+//		Package page for the given package.
+//	GET /<dir>
+//		Page listing packages under the given directory,
+//		assuming that there's no package with the given name.
+//	GET /<name>/<subpkg>
+//		Package page for the given subpackage.
+type Handler struct {
+	pkgs pathTree[*sallyPackage]
+}
 
-	pkgs := make([]packageInfo, 0, len(config.Packages))
+// CreateHandler builds a new handler
+// with the provided package configuration.
+func CreateHandler(config *Config) *Handler {
+	var pkgs pathTree[*sallyPackage]
 	for name, pkg := range config.Packages {
-		handler := newPackageHandler(config, name, pkg)
-		// Double-register so that "/foo"
-		// does not redirect to "/foo/" with a 300.
-		mux.Handle("/"+name, handler)
-		mux.Handle("/"+name+"/", handler)
+		baseURL := config.URL
+		if pkg.URL != "" {
+			// Package-specific override for the base URL.
+			baseURL = pkg.URL
+		}
+		modulePath := path.Join(baseURL, name)
+		docURL := "https://" + path.Join(config.Godoc.Host, modulePath)
 
-		pkgs = append(pkgs, packageInfo{
+		pkgs.Set(name, &sallyPackage{
 			Desc:       pkg.Desc,
-			ImportPath: handler.canonicalURL,
-			GitURL:     handler.gitURL,
-			GodocHome:  handler.godocHost + "/" + handler.canonicalURL,
+			ModulePath: modulePath,
+			DocURL:     docURL,
+			GitURL:     pkg.Repo,
 		})
 	}
-	sort.Slice(pkgs, func(i, j int) bool {
-		return pkgs[i].ImportPath < pkgs[j].ImportPath
-	})
-	mux.Handle("/", &indexHandler{pkgs: pkgs})
 
-	return mux
-}
-
-type indexHandler struct {
-	pkgs []packageInfo
-}
-
-type packageInfo struct {
-	Desc       string // package description
-	ImportPath string // canonical import path
-	GitURL     string // URL of the Git repository
-	GodocHome  string // documentation home URL
-}
-
-func (h *indexHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	// Index handler only supports '/'.
-	// ServeMux will call us for any '/foo' that is not a known package.
-	if r.Method != http.MethodGet || r.URL.Path != "/" {
-		http.NotFound(w, r)
-		return
-	}
-
-	data := struct{ Packages []packageInfo }{
-		Packages: h.pkgs,
-	}
-	if err := indexTemplate.Execute(w, data); err != nil {
-		http.Error(w, err.Error(), 500)
+	return &Handler{
+		pkgs: pkgs,
 	}
 }
 
-type packageHandler struct {
-	// Hostname of the godoc server, e.g. "godoc.org".
-	godocHost string
+var _ http.Handler = (*Handler)(nil)
 
-	// Name of the package relative to the vanity base URL.
-	// For example, "zap" for "go.uber.org/zap".
-	name string
-
-	// Path at which the Git repository is hosted.
-	// For example, "github.com/uber-go/zap".
-	gitURL string
-
+type sallyPackage struct {
 	// Canonical import path for the package.
-	canonicalURL string
+	ModulePath string
+
+	// Description of the package, if any.
+	Desc string
+
+	// URL at which documentation for the package can be found.
+	DocURL string
+
+	// URL at which the Git repository is hosted.
+	GitURL string
 }
 
-func newPackageHandler(cfg *Config, name string, pkg PackageConfig) *packageHandler {
-	baseURL := cfg.URL
-	if pkg.URL != "" {
-		baseURL = pkg.URL
-	}
-	canonicalURL := fmt.Sprintf("%s/%s", baseURL, name)
-
-	return &packageHandler{
-		godocHost:    cfg.Godoc.Host,
-		name:         name,
-		canonicalURL: canonicalURL,
-		gitURL:       pkg.Repo,
+func (h *Handler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+	if err := h.serveHTTP(w, r); err != nil {
+		if herr := new(httpError); errors.As(err, &herr) {
+			http.Error(w, herr.Message, herr.Code)
+		} else {
+			http.Error(w, err.Error(), 500)
+		}
 	}
 }
 
-func (h *packageHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+// httpError indicates that an HTTP error occurred.
+//
+// The caller will write the error code and message to the response.
+type httpError struct {
+	Code    int    // HTTP status code
+	Message string // error message
+}
+
+func httpErrorf(code int, format string, args ...interface{}) error {
+	return &httpError{
+		Code:    code,
+		Message: fmt.Sprintf(format, args...),
+	}
+}
+
+func (e *httpError) Error() string {
+	return fmt.Sprintf("status %d: %s", e.Code, e.Message)
+}
+
+// serveHTTP is similar to ServeHTTP, except it returns an error.
+//
+// If it returns an httpError,
+// the caller will write the error code and message to the response.
+func (h *Handler) serveHTTP(w http.ResponseWriter, r *http.Request) error {
 	if r.Method != http.MethodGet {
-		http.NotFound(w, r)
-		return
+		return httpErrorf(http.StatusNotFound, "method %q not allowed", r.Method)
 	}
 
-	// Extract the relative path to subpackages, if any.
-	//	"/foo/bar" => "/bar"
-	//	"/foo" => ""
-	relPath := strings.TrimPrefix(r.URL.Path, "/"+h.name)
+	path := strings.TrimPrefix(strings.TrimSuffix(r.URL.Path, "/"), "/")
 
-	data := struct {
-		Repo         string
-		CanonicalURL string
-		GodocURL     string
-	}{
-		Repo:         h.gitURL,
-		CanonicalURL: h.canonicalURL,
-		GodocURL:     fmt.Sprintf("https://%s/%s%s", h.godocHost, h.canonicalURL, relPath),
+	if pkg, suffix, ok := h.pkgs.Lookup(path); ok {
+		return h.servePackage(w, pkg, suffix)
 	}
-	if err := packageTemplate.Execute(w, data); err != nil {
-		http.Error(w, err.Error(), 500)
+	return h.serveIndex(w, path, h.pkgs.ListByPath(path))
+}
+
+func (h *Handler) servePackage(w http.ResponseWriter, pkg *sallyPackage, suffix string) error {
+	return packageTemplate.Execute(w,
+		struct {
+			ModulePath string
+			GitURL     string
+			DocURL     string
+		}{
+			ModulePath: pkg.ModulePath,
+			GitURL:     pkg.GitURL,
+			DocURL:     pkg.DocURL + suffix,
+		})
+}
+
+func (h *Handler) serveIndex(w http.ResponseWriter, path string, pkgs []*sallyPackage) error {
+	if len(pkgs) == 0 {
+		return httpErrorf(http.StatusNotFound, "no packages found under path: %s", path)
 	}
+
+	return indexTemplate.Execute(w,
+		struct{ Packages []*sallyPackage }{
+			Packages: pkgs,
+		})
 }