+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Generate a package listing for sub-paths
+  that match a subset of the known packages.
+
 ## [1.4.0]
 ### Added
 - Publish a Docker image to GitHub Container Registry.